requirements/development.txt
-r base.txt

# Development tools
django-debug-toolbar>=4.2,<5.0
ipython>=8.0,<9.0

# Testing
pytest>=8.0,<9.0
pytest-django>=4.7,<5.0
pytest-cov>=4.1,<6.0
factory-boy>=3.3,<4.0

# Code quality
black>=24.0,<25.0
isort>=5.13,<6.0
flake8>=7.0,<8.0

# Type checking
mypy>=1.8,<2.0
django-stubs>=4.2,<5.0
djangorestframework-stubs>=3.14,<4.0
